Implement coercion for Vec<T>
This implements the two-step Helix coercion protocol for Vec<T> from
Ruby Arrays.

The implementation uses the constraints that UncheckedValue<T> must be
implemented for VALUE and ToRust<T> must be implemented for
CheckedValue<T>. This ensures that the elements in the Vec can be
successfully coerced into the right type, and we let the elements handle
the coercion themselves, while the implementation for Vec<T> just
handles creating the Vec and filling it with the coerced elements.

This has the limitation that the Array passed on from Ruby needs to
homogenous, like Rust Vecs.